Добавление методов
