+++ pom.xml
@@ -29,6 +29,14 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-devtools</artifactId>
+			<scope>runtime</scope>
+			<optional>true</optional>
+		</dependency>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
@@ -98,6 +106,13 @@
 <!--			<artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>-->
 <!--			<version>2.6.0</version>-->
 <!--		</dependency>-->
+
+		<!--        отображение страниц -->
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-thymeleaf</artifactId>
+		</dependency>
+
 		<!--        тесты -->
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
