--- src/main/java/ru/kpepskot/sport_shop/dto/UserInitUpdateDto.java
@@ -1,4 +1,0 @@
-package ru.kpepskot.sport_shop.dto;
-
-public class UserInitUpdateDto {
-}
