+++ src/main/java/ru/kpepskot/sport_shop/dto/user/UserDtoMapper.java
@@ -1,4 +1,4 @@
-package ru.kpepskot.sport_shop.dto;
+package ru.kpepskot.sport_shop.dto.user;
 
 import org.springframework.stereotype.Component;
 import ru.kpepskot.sport_shop.entity.User;
@@ -7,7 +7,7 @@ import ru.kpepskot.sport_shop.entity.User;
 @Component
 public class UserDtoMapper {
 
-    public UserDto UserToUserDto(User user) {
+    public UserDto userToUserDto(User user) {
         return UserDto.builder()
                 .id(user.getId())
                 .userEmail(user.getUserEmail())
