+++ src/main/java/ru/kpepskot/sport_shop/dto/user/UserInitDtoMapper.java
@@ -1,4 +1,4 @@
-package ru.kpepskot.sport_shop.dto;
+package ru.kpepskot.sport_shop.dto.user;
 
 import org.springframework.stereotype.Component;
 import ru.kpepskot.sport_shop.entity.User;
