+++ src/main/java/ru/kpepskot/sport_shop/dto/user/UserInitUpdateDto.java
@@ -1,0 +1,4 @@
+package ru.kpepskot.sport_shop.dto.user;
+
+public class UserInitUpdateDto {
+}
