+++ src/main/java/ru/kpepskot/sport_shop/service/impl/UserServiceImpl.java
@@ -1,15 +1,18 @@
-package ru.kpepskot.sport_shop.service;
+package ru.kpepskot.sport_shop.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.kpepskot.sport_shop.constant.Role;
-import ru.kpepskot.sport_shop.dto.*;
+import ru.kpepskot.sport_shop.dto.user.*;
 import ru.kpepskot.sport_shop.entity.User;
 import ru.kpepskot.sport_shop.error.NotFoundException;
 import ru.kpepskot.sport_shop.repository.UserRepository;
+import ru.kpepskot.sport_shop.service.UserService;
 
 import javax.transaction.Transactional;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -23,7 +26,7 @@ public class UserServiceImpl implements UserService {
     public UserDto findUserById(Long id) {
         Optional<User> optionalUser = userRepository.findById(id);
         if (optionalUser.isPresent()) {
-            return userDtoMapper.UserToUserDto(optionalUser.get());
+            return userDtoMapper.userToUserDto(optionalUser.get());
         } else {
             throw new NotFoundException("Пользователь с id = " + id + " не был найден");
         }
@@ -34,7 +37,7 @@ public class UserServiceImpl implements UserService {
     public UserDto createUser(UserInitDto userInitDto) {
         User user = userInitDtoMapper.UserInitDtoToUser(userInitDto);
         user.setUserRole(Role.USER.toString());
-        return userDtoMapper.UserToUserDto(userRepository.save(user));
+        return userDtoMapper.userToUserDto(userRepository.save(user));
     }
 
     @Override
@@ -48,5 +51,10 @@ public class UserServiceImpl implements UserService {
     public void deleteUserById(Long userId) {
        findUserById(userId);
        userRepository.deleteById(userId);
+    }
+
+    @Override
+    public List<UserDto> findAllUsers() {
+        return userRepository.findAll().stream().map(userDtoMapper::userToUserDto).collect(Collectors.toList());
     }
 }
