+++ src/main/resources/static/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Добро пожаловать в &ndash; pet-проект для отработки навыков!</title>
+</head>
+<body>
+    <h1>Добро пожаловать!</h1>
+</body>
+</html>
